Reorg from utils to test_cli, speed fix for cli tests
The speed fix basically does the following:
* mock main.command to listen to function calls
* stop calling quilt <args> in a subprocess (quilt has slow load time
currently, see #194)
* Generate parser via main.argument_parser
* call main.main() with parser args

Also, some test_cli related objects were moved from test/util.py to
test/test_cli.py
