+++ quilt/test/test_cli.py
@@ -56,12 +56,14 @@ removal (or incompatible additions) on the API side.
 import os
 import sys
 import json
+import inspect
+import collections
 from subprocess import call, check_output, PIPE, CalledProcessError
 
 import pytest
 
-from .utils import BasicQuiltTestCase, ArgparseIntrospector, RecursiveMappingWrapper, PACKAGE_DIR
-from ..tools import command
+from .utils import BasicQuiltTestCase
+from ..tools import command, main
 
 # inspect.argspec is deprecated, so
 try:
@@ -69,16 +71,29 @@ try:
 except ImportError:
     from inspect import signature
 
+
+## "Static" vars
+_TEST_DIR = os.path.abspath(inspect.stack()[0][1])
+PACKAGE_DIR = os.path.join('/', *_TEST_DIR.split(os.path.sep)[:-3])
+
 # When a test for CLI params is made, append the param key paths that
 # the test addresses to this variable.
-# Example key path by calling get_all_param_paths()
+# Get an example key path by calling get_all_param_paths()
 TESTED_PARAMS = []
 
+
+## KNOWN_PARAMS
 # When adding a new param to the cli, add the param here.
 # New or missing cli params can be found in test errors as keypaths,
 # which are simply lists of dict keys.  These can be added to or
 # removed from KNOWN_PARAMS as befits your situation.
 # The end value should be the argparse 'dest' arg (if present, or None.
+
+#TODO: KNOWN_PARAMS could potentially be a path list instead of a dict tree.
+#      This has the benefit of being easier to understand at a glance, and of
+#      being easy to update from failed test copypasta.
+#      A nested list wouldn't really do -- this needs to represent both
+#      positional and keyword arguments.
 KNOWN_PARAMS = {
     0: {
         "access": {
@@ -169,6 +184,8 @@ KNOWN_PARAMS = {
 
 #TODO: Fix ambiguity of positional options vs anonymous positionals with separate options following them
 
+
+## Helper functions
 def get_all_param_paths():
     """Get all paths by introspecting the ArgumentParser from main()
 
@@ -192,10 +209,9 @@ def get_current_param_tree():
     This has the same structure as KNOWN_PARAMS, but contains the
     current params from main().
 
-    :rtype: dict
+    :rtype: ArgparseIntrospector
     """
-    from ..tools.main import argument_parser
-    return ArgparseIntrospector(argument_parser()).as_dict()
+    return ArgparseIntrospector(main.argument_parser())
 
 
 def get_missing_key_paths(a, b, exhaustive=False):
@@ -261,18 +277,257 @@ def fails_binding(func, args, kwargs):
         return error.args[0] if error.args and error.args[0] else "Failed to bind"
 
 
-class MockCommand(object):
-    _target = command
+class ArgparseIntrospector(collections.Mapping):
+    def __init__(self, argparse_object, ignore=('-h', '--help'), leaf_func=lambda x: getattr(x, 'dest', None)):
+        """Create a parameter tree from an argparse parser.
+
+        Once an ArgumentParser is configured, it can be passed as the param
+        `argparse_object`, which makes the parameters available as a tree.
+
+        If any parameter has an item from `ignore` in its opton strings, it
+        will be excluded from the tree and from iteration.
+
+        The `leaf_func` param warrants a bit more of an in-depth description.
+        When an object is found not to have any further children, it is passed
+        to the leaf_func.  The returned value is then entered into the tree.
+        The default function simply returns the object's "dest" attribute,
+        which acts as a very short description of what the argument is, in
+        this case.
+
+        :param argparse_object: A (configured) ArgumentParser object
+        :param ignore: Ignore arguments in given iterable (-h, --help by default)
+        :param leaf_func: Receives an object.  Data returned is entered into tree.
+        """
+        self.data = argparse_object
+        self.ignored = ignore
+        self.leaf_func = leaf_func
+
+    def __getitem__(self, k):
+        if k in self.ignored:
+            raise KeyError("Ignored: {!r}".format(k))
+
+        if isinstance(k, int):
+            positionals = self.data._get_positional_actions()
+            try:
+                action = positionals[k]
+            except IndexError:
+                raise KeyError(k)
+            choices = action.choices
+            if not choices:
+                return self.leaf_func(action)
+            return collections.OrderedDict((choice, ArgparseIntrospector(parser))
+                                            for choice, parser in choices.items())
+        else:
+            optionals = self.data._get_optional_actions()
+            for o in optionals:
+                if k in o.option_strings:
+                    return self.leaf_func(o)
+        raise KeyError(k)
+
+    def __len__(self):
+        return len(tuple(self.__iter__()))
+
+    def __iter__(self):
+        for index, positional in enumerate(self.data._get_positional_actions()):
+            yield index
+        for optional in self.data._get_optional_actions():
+            if any(s in self.ignored for s in optional.option_strings):
+                continue
+            yield optional.option_strings[0]
+
+    def as_dict(self):
+        def _mcopy(old):
+            if not old:
+                return self.leaf_func(old.data)
+            if not isinstance(old, collections.Mapping):
+                return old
+            new = {}
+            for k, v in old.items():
+                new[k] = _mcopy(v) if v else None
+            return new
+        return _mcopy(self)
+
+    def __repr__(self):
+        args = '\n'.join(repr(x) for x in self)
+        return 'ArgparseIntrospector with args:\n' + args
+
+
+class RecursiveMappingWrapper(collections.MutableMapping):
+    """Wrap a mapping, providing recursive access to items via key lists
+
+    allows:
+    >>> x = RecursiveMappingWrapper({'a': {'b': 'c'}})
+    >>> x[['a', 'b']]
+    'c'
+    >>> x[['a', 'n']] = 9
+    >>> x.obj
+    {'a': {'b': 'c', 'n': 9}}
+
+    It does not implicitly add new dicts/mappings by default, so:
+    >>> x = RecursiveMappingWrapper({})
+    >>> x[['foo', 'bar']] = 'baz'
+    <raises KeyError('Invalid key path', ['foo'])>
+    >>> x['foo'] = {}
+    >>> x[['foo', 'bar']] = 'baz'
+    >>> x.obj
+    {'foo': {'bar': 'baz'}}
+
+    However, if `implicit_fill` is truthy, then the following occurs:
+    >>> x = RecursiveMappingWrapper({}, implicit_fill=True)
+    >>> x[['foo', 'bar']]
+    <raises KeyError('Invalid key path', ['foo'])>
+    >>> x[['foo', 'bar']] = 'baz'
+    >>> x.obj
+    {'foo': {'bar': 'baz'}}
+
+    :param obj: mapping/dict to wrap
+    :param implicit_fill: implicitly resolve interim keys to new dict objects
+    """
+    def __init__(self, obj, implicit_fill=False):
+        if not isinstance(obj, collections.Mapping):
+            raise TypeError("`obj` must be a mapping.")
+        super(RecursiveMappingWrapper, self).__init__()
+        self.obj = obj
+        self.implicit_fill = implicit_fill
+
+    def __delitem__(self, key):
+        """Delete by key or key path."""
+        if not isinstance(key, list):
+            del self.obj[key]
+            return
+        keys = key
+        try:
+            key = keys[-1]
+        except IndexError:
+            raise KeyError("empty key list")
+        obj = self._get_next_to_last(keys)
+        try:
+            del obj[key]
+        except KeyError:
+            raise KeyError("Invalid key path", keys)
+
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.obj == other.obj
+        return self.obj == other
+
+    def __getitem__(self, key):
+        """Fetch value by key or key path."""
+        if not isinstance(key, list):
+            return self.obj[key]
+        keys = key
+        try:
+            key = keys[-1]
+        except IndexError:
+            raise KeyError("empty key list")
+        obj = self._get_next_to_last(keys)
+        try:
+            return obj[key]
+        except KeyError:
+            raise KeyError("Invalid key path", keys)
+
+    def __iter__(self):
+        """Iterate over keys (not key paths)."""
+        return iter(self.obj)
+
+    def __len__(self):
+        return len(self.obj)
+
+    def __repr__(self):
+        return "{}({})".format(type(self).__name__, self.obj)
+
+    def __setitem__(self, key, value):
+        if not isinstance(key, list):
+            self.obj[key] = value
+            return
+        keys = key
+        try:
+            key = keys[-1]
+        except IndexError:
+            raise KeyError("empty key list")
+        obj = self._get_next_to_last(keys, implicit_fill=self.implicit_fill)
+        obj[key] = value
+
+    def _get_next_to_last(self, keys, implicit_fill=False):
+        # list is our recursive type, because it is mutable, and can't be
+        # used for a key, anyways.
+        if not isinstance(keys, list):
+            raise TypeError('`keys` must be a list')
+        obj = self.obj
+        path = []
+        try:
+            for k in keys[:-1]:
+                path.append(k)
+                if implicit_fill:
+                    obj[k] = obj.get(k, {})
+                obj = obj[k]
+        except KeyError:
+            raise KeyError("Invalid key path", path)
+        return obj
+
+    def __getattr__(self, item):
+        return getattr(self.obj, item)
+
+    def iterpaths(self, sortkey=lambda x: repr(x)):
+        """Iterate recursively over contained keys and key paths"""
+        if sortkey:
+            keys = sorted(self, key=sortkey)
+        else:
+            keys = self.keys()
+        for key in keys:
+            value = self[key]
+            yield [key]
+            if isinstance(value, collections.Mapping):
+                if not isinstance(value, self.__class__):
+                    value = self.__class__(value)
+                for path in value.iterpaths():
+                    yield [key] + path
+
+
+class MockObject(object):
+    """Mocks objects with callables.
+
+    For any attribute that is accessed, it gives a function that prints
+    JSON information to stdout instead.
+    """
+    _target = None
+    __result = {}
+
+    def __init__(self, target):
+        self._target = target
+
+    @property
+    def _result(self):
+        """Deletes itself after being read."""
+        result = self.__result
+        self.__result = {}
+        return result
+
+    @_result.setter
+    def _result(self, v):
+        self.__result = v
+
     def __getattr__(self, funcname):
         def dummy_func(*args, **kwargs):
-            result = {'func': funcname, 'matched': hasattr(self._target, funcname), 'args': args, 'kwargs': kwargs}
+            result = {
+                'func': funcname,
+                'matched': hasattr(self._target, funcname),
+                'args': args,
+                'kwargs': kwargs
+                }
             print(json.dumps(result, indent=4))
+            self._result = result
         return dummy_func
 
 
 class TestCLI(BasicQuiltTestCase):
     def setUp(self):
+        self.mock_command = MockObject(command)
+        main.command = self.mock_command
         self.param_tree = get_current_param_tree()
+        self.parser = self.param_tree.data
+
+        # if using subprocess calls
         self.env = os.environ.copy()
         self.env['QUILT_CLI_TEST'] = "True"
         self.env['PYTHON_PATH'] = PACKAGE_DIR
@@ -280,7 +535,28 @@ class TestCLI(BasicQuiltTestCase):
         self.quilt_command = [sys.executable, '-c', 'from quilt.tools import main; main.main()']
 
     def tearDown(self):
-        pass
+        main.command = self.mock_command._target
+
+    def execute(self, cli_args):
+
+        #TODO: Enable switching between direct and subprocess modes
+
+        # Subprocess mode -- snippets, untested, may not work
+        # quilt = self.quilt_command
+        # env = self.env
+        # cmd = quilt + ['config']
+        # result = json.loads(check_output(cmd, env=env).decode())  # Terminal default decoding
+
+        # direct mode -- uses quilt.tools.main.main()
+        result = {}
+        try:
+            main.main(cli_args)
+        except SystemExit as error:
+            result['return code'] = error.args[0] if error.args else 0
+        else:
+            result['return code'] = 0
+        result.update(self.mock_command._result)
+        return result
 
     def test_cli_new_param(self):
         missing_paths = get_missing_key_paths(self.param_tree, KNOWN_PARAMS, exhaustive=True)
@@ -295,10 +571,7 @@ class TestCLI(BasicQuiltTestCase):
             pytest.fail(message.format('\n\t'.join(repr(x) for x in missing_paths)))
 
     def test_cli_command_config(self):
-        """Ensures the 'cli' command calls a specific API"""
-        quilt = self.quilt_command
-        env = self.env
-
+        """Ensures the 'config' command calls a specific API"""
         ## This test covers the following arguments that require testing
         TESTED_PARAMS.extend([
             [0, 'config'],
@@ -309,13 +582,14 @@ class TestCLI(BasicQuiltTestCase):
             'config --badparam'.split(),
             ]
         for args in expect_fail_2_args:
-            assert call(quilt + args, env=env) == 2
+            assert self.execute(args)['return code'] == 2
 
         ## This section tests for appropriate types and values.
-        cmd = quilt + ['config']
-        result = json.loads(check_output(cmd, env=env).decode())  # Terminal default decoding
+        cmd = ['config']
+        result = self.execute(cmd)
 
         # General tests
+        assert result['return code'] == 0
         assert result['matched'] is True  # func name recognized by MockCommand class?
         func = getattr(command, result['func'])
         assert not fails_binding(func, args=result['args'], kwargs=result['kwargs'])
@@ -326,9 +600,6 @@ class TestCLI(BasicQuiltTestCase):
         assert not result['kwargs']
 
     def test_cli_command_push(self):
-        quilt = self.quilt_command
-        env = self.env
-
         ## This test covers the following arguments that require testing
         TESTED_PARAMS.extend([
             [0, 'push'],
@@ -345,11 +616,11 @@ class TestCLI(BasicQuiltTestCase):
             'push --public --reupload'.split(),
             ]
         for args in expect_fail_2_args:
-            assert call(quilt + args, env=env) == 2
+            assert self.execute(args)['return code'] == 2
 
         ## This section tests for appropriate types and values.
-        cmd = quilt + 'push fakeuser/fakepackage'.split()
-        result = json.loads(check_output(cmd, env=env).decode())  # Terminal default decoding
+        cmd = 'push fakeuser/fakepackage'.split()
+        result = self.execute(cmd)
 
         # General tests
         assert result['matched'] is True    # func name recognized by MockCommand class?
@@ -364,8 +635,8 @@ class TestCLI(BasicQuiltTestCase):
         assert result['kwargs']['package'] == 'fakeuser/fakepackage'
 
         ## Test the flags as well..
-        cmd = quilt + 'push --reupload --public fakeuser/fakepackage'.split()
-        result = json.loads(check_output(cmd, env=env).decode())  # Terminal default decoding
+        cmd = 'push --reupload --public fakeuser/fakepackage'.split()
+        result = self.execute(cmd)
 
         # General tests
         assert result['matched'] is True    # func name recognized by MockCommand class?
